moved the array tricks functions into a seperate file
